requirements.txt
python-telegram-bot==20.4
instagrapi==1.18.0
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.1